결제로직 살짝 수정 및 주석 추가

+++ src/main/java/com/shop/domain/OrderItem.java
@@ -15,7 +15,7 @@ public class OrderItem {
     private Long productOptionNo; 
     
     private Integer quantity;         
-    private Integer unitPrice;      
+    private Long unitPrice;      
     
     private String itemName;      
     private String itemSize;
